refactor: extract shared SPOCP parser to pkg/spocputil
Move the duplicated advancedParser implementation from both
pkg/issuance/parser.go and pkg/httphelpers/middleware_jwt.go
into a new shared pkg/spocputil package.

This eliminates 19.8% new-code duplication that was failing the
SonarCloud quality gate (threshold: 3%).
