+++ pkg/httphelpers/middleware_jwt.go
@@ -1,7 +1,6 @@
 package httphelpers
 
 import (
-	"bufio"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -11,9 +10,9 @@ import (
 	"strings"
 	"sync"
 	"time"
-	"unicode"
 
 	"github.com/SUNET/vc/pkg/model"
+	"github.com/SUNET/vc/pkg/spocputil"
 
 	"github.com/gin-gonic/gin"
 	"github.com/lestrrat-go/jwx/v3/jwk"
@@ -400,7 +399,7 @@ func BuildSPOCPEngine(cfg model.APIAuth) (*SafeEngine, error) {
 	engine := spocp.New()
 
 	for i, r := range cfg.Rules {
-		elem, err := parseAdvancedSExp(r)
+		elem, err := spocputil.ParseAdvancedSExp(r)
 		if err != nil {
 			return nil, fmt.Errorf("invalid inline SPOCP rule #%d: %w", i+1, err)
 		}
@@ -411,262 +410,12 @@ func BuildSPOCPEngine(cfg model.APIAuth) (*SafeEngine, error) {
 	}
 
 	if hasFile {
-		if err := loadRulesFromFile(engine, cfg.RulesFile); err != nil {
+		if err := spocputil.LoadRulesFromFile(engine, cfg.RulesFile); err != nil {
 			return nil, fmt.Errorf("failed to load SPOCP rules from %s: %w", cfg.RulesFile, err)
 		}
 	}
 
 	return &SafeEngine{engine: engine}, nil
-}
-
-// parseRulesFile parses a rules file and returns the elements (best-effort)
-func parseRulesFile(path string) []sexp.Element {
-	f, err := os.Open(filepath.Clean(path))
-	if err != nil {
-		return nil
-	}
-	defer f.Close()
-
-	var elems []sexp.Element
-	scanner := bufio.NewScanner(f)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		elem, err := parseAdvancedSExp(line)
-		if err != nil {
-			continue
-		}
-		elems = append(elems, elem)
-	}
-	return elems
-}
-
-// loadRulesFromFile reads human-readable SPOCP rules (one per line) from a file
-// and adds them to the engine using parseAdvancedSExp.
-func loadRulesFromFile(engine *spocp.AdaptiveEngine, path string) error {
-	f, err := os.Open(filepath.Clean(path))
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	scanner := bufio.NewScanner(f)
-	lineNum := 0
-	for scanner.Scan() {
-		lineNum++
-		line := strings.TrimSpace(scanner.Text())
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue // skip blanks and comments
-		}
-		elem, err := parseAdvancedSExp(line)
-		if err != nil {
-			return fmt.Errorf("line %d: %w", lineNum, err)
-		}
-		if err := validateRuleElement(elem, fmt.Sprintf("file line %d", lineNum)); err != nil {
-			return err
-		}
-		engine.AddRuleElement(elem)
-	}
-	return scanner.Err()
-}
-
-// parseAdvancedSExp parses a human-readable ("advanced form") S-expression into
-// a sexp.Element. This allows users to write rules as:
-//
-//	(api (method POST)(path /api/v1/upload)(subject alice))
-//
-// instead of canonical form:
-//
-//	(3:api(6:method4:POST)(4:path15:/api/v1/upload)(7:subject5:alice))
-//
-// It also supports star forms:
-//
-//	(*)                        → wildcard
-//	(* prefix /api/v1/)        → prefix match
-//	(* suffix .pdf)            → suffix match
-//	(* set read write delete)  → set match
-func parseAdvancedSExp(input string) (sexp.Element, error) {
-	input = strings.TrimSpace(input)
-	if input == "" {
-		return nil, fmt.Errorf("empty S-expression")
-	}
-
-	p := &advancedParser{input: []rune(input), pos: 0}
-	elem, err := p.parse()
-	if err != nil {
-		return nil, err
-	}
-
-	// Ensure we consumed all input.
-	p.skipWhitespace()
-	if p.pos < len(p.input) {
-		return nil, fmt.Errorf("unexpected trailing input at position %d", p.pos)
-	}
-	return elem, nil
-}
-
-type advancedParser struct {
-	input []rune
-	pos   int
-}
-
-func (p *advancedParser) skipWhitespace() {
-	for p.pos < len(p.input) && unicode.IsSpace(p.input[p.pos]) {
-		p.pos++
-	}
-}
-
-func (p *advancedParser) parse() (sexp.Element, error) {
-	p.skipWhitespace()
-	if p.pos >= len(p.input) {
-		return nil, fmt.Errorf("unexpected end of input")
-	}
-
-	if p.input[p.pos] == '(' {
-		return p.parseList()
-	}
-	return p.parseAtom()
-}
-
-func (p *advancedParser) parseAtom() (*sexp.Atom, error) {
-	p.skipWhitespace()
-	if p.pos >= len(p.input) {
-		return nil, fmt.Errorf("unexpected end of input, expected atom")
-	}
-
-	start := p.pos
-	for p.pos < len(p.input) && p.input[p.pos] != '(' && p.input[p.pos] != ')' && !unicode.IsSpace(p.input[p.pos]) {
-		p.pos++
-	}
-	if p.pos == start {
-		return nil, fmt.Errorf("empty atom at position %d", start)
-	}
-	return sexp.NewAtom(string(p.input[start:p.pos])), nil
-}
-
-func (p *advancedParser) parseList() (sexp.Element, error) {
-	if p.input[p.pos] != '(' {
-		return nil, fmt.Errorf("expected '(' at position %d", p.pos)
-	}
-	p.pos++ // skip '('
-
-	p.skipWhitespace()
-	if p.pos >= len(p.input) {
-		return nil, fmt.Errorf("unclosed '('")
-	}
-
-	// Parse the tag atom.
-	tag, err := p.parseAtom()
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse list tag: %w", err)
-	}
-
-	// Handle star forms: tag == "*"
-	if tag.Value == "*" {
-		return p.parseStarForm()
-	}
-
-	// Parse remaining elements until ')'.
-	var elements []sexp.Element
-	for {
-		p.skipWhitespace()
-		if p.pos >= len(p.input) {
-			return nil, fmt.Errorf("unclosed list '%s'", tag.Value)
-		}
-		if p.input[p.pos] == ')' {
-			p.pos++ // skip ')'
-			return sexp.NewList(tag.Value, elements...), nil
-		}
-		elem, err := p.parse()
-		if err != nil {
-			return nil, err
-		}
-		// Treat a bare * inside a tagged list as a wildcard star form,
-		// so (method *) is equivalent to (method (*)).
-		// Treat a trailing * (e.g. /api/v1/*) as a prefix star form,
-		// so (path /api/v1/*) is equivalent to (path (* prefix /api/v1/)).
-		if atom, ok := elem.(*sexp.Atom); ok {
-			if atom.Value == "*" {
-				elem = &starform.Wildcard{}
-			} else if before, found := strings.CutSuffix(atom.Value, "*"); found {
-				elem = &starform.Prefix{Value: before}
-			}
-		}
-		elements = append(elements, elem)
-	}
-}
-
-// parseStarForm parses the body of a star form after the '*' tag.
-// At this point the opening '(' and '*' have been consumed.
-func (p *advancedParser) parseStarForm() (sexp.Element, error) {
-	p.skipWhitespace()
-	if p.pos >= len(p.input) {
-		return nil, fmt.Errorf("unclosed star form")
-	}
-
-	// (*) → wildcard
-	if p.input[p.pos] == ')' {
-		p.pos++
-		return &starform.Wildcard{}, nil
-	}
-
-	// Read the star form type: set, prefix, suffix, range...
-	formType, err := p.parseAtom()
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse star form type: %w", err)
-	}
-
-	switch formType.Value {
-	case "prefix":
-		return p.parsePrefixSuffix(true)
-	case "suffix":
-		return p.parsePrefixSuffix(false)
-	case "set":
-		return p.parseSet()
-	default:
-		return nil, fmt.Errorf("unsupported star form type %q", formType.Value)
-	}
-}
-
-func (p *advancedParser) parsePrefixSuffix(isPrefix bool) (sexp.Element, error) {
-	p.skipWhitespace()
-	value, err := p.parseAtom()
-	if err != nil {
-		return nil, fmt.Errorf("expected value for prefix/suffix star form: %w", err)
-	}
-	p.skipWhitespace()
-	if p.pos >= len(p.input) || p.input[p.pos] != ')' {
-		return nil, fmt.Errorf("expected ')' to close prefix/suffix star form")
-	}
-	p.pos++
-	if isPrefix {
-		return &starform.Prefix{Value: value.Value}, nil
-	}
-	return &starform.Suffix{Value: value.Value}, nil
-}
-
-func (p *advancedParser) parseSet() (sexp.Element, error) {
-	var elements []sexp.Element
-	for {
-		p.skipWhitespace()
-		if p.pos >= len(p.input) {
-			return nil, fmt.Errorf("unclosed set star form")
-		}
-		if p.input[p.pos] == ')' {
-			p.pos++
-			if len(elements) == 0 {
-				return nil, fmt.Errorf("empty set star form")
-			}
-			return &starform.Set{Elements: elements}, nil
-		}
-		elem, err := p.parse()
-		if err != nil {
-			return nil, err
-		}
-		elements = append(elements, elem)
-	}
 }
 
 // extractSPOCPSubject returns the identity to use as the SPOCP subject
@@ -682,8 +431,8 @@ func extractSPOCPSubject(token jwt.Token) string {
 	return ""
 }
 
-// BuildSPOCPQuery constructs a SPOCP query S-expression for the current HTTP
-// request, including service, method, path, subject, authentic source and scope:
+// buildSPOCPQuery constructs a SPOCP query S-expression for the current HTTP
+// request, service name, and JWT subject:
 //
 //	(vc (service apigw)(method POST)(path /api/v1/upload)(subject alice@sunet.se)(authentic_source SUNET)(scope eduid))
 //
